requirements.txt
python-dotenv>=0.20.0
selenium>=4.1.3
yaml>=6.0